Fix audio: generated speech plays when received
